Split up MessagePropertyTextMap
